Create RedirectLoopDetector
