+++ tests/ContainerTest.php
@@ -373,301 +373,8 @@ class ContainerTest extends TestCase
         $this->assertCount(0, $this->container);
     }
 
-    /**
-     * @dataProvider hasRedirectLoopDataProvider
-     *
-     * @param array<int, array<int, RequestInterface|ResponseInterface>> $httpTransactions
-     * @param bool $expectedHasRedirectLoop
-     */
-    public function testHasRedirectLoop(array $httpTransactions, bool $expectedHasRedirectLoop)
-    {
-        foreach ($httpTransactions as $httpTransaction) {
-            $this->container[] = $httpTransaction;
-        }
-
-        $this->assertEquals($expectedHasRedirectLoop, $this->container->hasRedirectLoop());
-    }
-
-    public function hasRedirectLoopDataProvider(): array
-    {
-        return [
-            'single 200 response' => [
-                'httpTransactions' => [
-                    [
-                        HttpTransaction::KEY_REQUEST => \Mockery::mock(RequestInterface::class),
-                        HttpTransaction::KEY_RESPONSE => $this->createResponse(200),
-                        HttpTransaction::KEY_ERROR => null,
-                        HttpTransaction::KEY_OPTIONS => []
-                    ],
-                ],
-                'expectedHasRedirectLoop' => false,
-            ],
-            'contains non-redirect response (200)' => [
-                'httpTransactions' => [
-                    [
-                        HttpTransaction::KEY_REQUEST => \Mockery::mock(RequestInterface::class),
-                        HttpTransaction::KEY_RESPONSE => $this->createResponse(301),
-                        HttpTransaction::KEY_ERROR => null,
-                        HttpTransaction::KEY_OPTIONS => []
-                    ],
-                    [
-                        HttpTransaction::KEY_REQUEST => \Mockery::mock(RequestInterface::class),
-                        HttpTransaction::KEY_RESPONSE => $this->createResponse(200),
-                        HttpTransaction::KEY_ERROR => null,
-                        HttpTransaction::KEY_OPTIONS => []
-                    ],
-                ],
-                'expectedHasRedirectLoop' => false,
-            ],
-            'contains non-redirect response (404)' => [
-                'httpTransactions' => [
-                    [
-                        HttpTransaction::KEY_REQUEST => \Mockery::mock(RequestInterface::class),
-                        HttpTransaction::KEY_RESPONSE => $this->createResponse(301),
-                        HttpTransaction::KEY_ERROR => null,
-                        HttpTransaction::KEY_OPTIONS => []
-                    ],
-                    [
-                        HttpTransaction::KEY_REQUEST => \Mockery::mock(RequestInterface::class),
-                        HttpTransaction::KEY_RESPONSE => $this->createResponse(404),
-                        HttpTransaction::KEY_ERROR => null,
-                        HttpTransaction::KEY_OPTIONS => []
-                    ],
-                ],
-                'expectedHasRedirectLoop' => false,
-            ],
-            'only redirects, no loop (all different)' => [
-                'httpTransactions' => [
-                    [
-                        HttpTransaction::KEY_REQUEST => $this->createRequest('GET', 'http://example.com/'),
-                        HttpTransaction::KEY_RESPONSE => $this->createResponse(301),
-                        HttpTransaction::KEY_ERROR => null,
-                        HttpTransaction::KEY_OPTIONS => []
-                    ],
-                    [
-                        HttpTransaction::KEY_REQUEST => $this->createRequest('GET', 'http://example.com/1'),
-                        HttpTransaction::KEY_RESPONSE => $this->createResponse(301),
-                        HttpTransaction::KEY_ERROR => null,
-                        HttpTransaction::KEY_OPTIONS => []
-                    ],
-                ],
-                'expectedHasRedirectLoop' => false,
-            ],
-            'method change within apparent loop is not loop' => [
-                'httpTransactions' => [
-                    [
-                        HttpTransaction::KEY_REQUEST => $this->createRequest('HEAD', 'http://example.com/'),
-                        HttpTransaction::KEY_RESPONSE => $this->createResponse(301),
-                        HttpTransaction::KEY_ERROR => null,
-                        HttpTransaction::KEY_OPTIONS => []
-                    ],
-                    [
-                        HttpTransaction::KEY_REQUEST => $this->createRequest('GET', 'http://example.com/'),
-                        HttpTransaction::KEY_RESPONSE => $this->createResponse(301),
-                        HttpTransaction::KEY_ERROR => null,
-                        HttpTransaction::KEY_OPTIONS => []
-                    ],
-                ],
-                'expectedHasRedirectLoop' => false,
-            ],
-            'redirecting directly back to self' => [
-                'httpTransactions' => [
-                    [
-                        HttpTransaction::KEY_REQUEST => $this->createRequest('GET', 'http://example.com/'),
-                        HttpTransaction::KEY_RESPONSE => $this->createResponse(301),
-                        HttpTransaction::KEY_ERROR => null,
-                        HttpTransaction::KEY_OPTIONS => []
-                    ],
-                    [
-                        HttpTransaction::KEY_REQUEST => $this->createRequest('GET', 'http://example.com/'),
-                        HttpTransaction::KEY_RESPONSE => $this->createResponse(301),
-                        HttpTransaction::KEY_ERROR => null,
-                        HttpTransaction::KEY_OPTIONS => []
-                    ],
-                ],
-                'expectedHasRedirectLoop' => true,
-            ],
-            'redirecting indirectly back to self' => [
-                'httpTransactions' => [
-                    [
-                        HttpTransaction::KEY_REQUEST => $this->createRequest('GET', 'http://example.com/'),
-                        HttpTransaction::KEY_RESPONSE => $this->createResponse(301),
-                        HttpTransaction::KEY_ERROR => null,
-                        HttpTransaction::KEY_OPTIONS => []
-                    ],
-                    [
-                        HttpTransaction::KEY_REQUEST => $this->createRequest('GET', 'http://example.com/1'),
-                        HttpTransaction::KEY_RESPONSE => $this->createResponse(301),
-                        HttpTransaction::KEY_ERROR => null,
-                        HttpTransaction::KEY_OPTIONS => []
-                    ],
-                    [
-                        HttpTransaction::KEY_REQUEST => $this->createRequest('GET', 'http://example.com/'),
-                        HttpTransaction::KEY_RESPONSE => $this->createResponse(301),
-                        HttpTransaction::KEY_ERROR => null,
-                        HttpTransaction::KEY_OPTIONS => []
-                    ],
-                ],
-                'expectedHasRedirectLoop' => true,
-            ],
-            'redirecting indirectly back to self (with method group change)' => [
-                'httpTransactions' => [
-                    [
-                        HttpTransaction::KEY_REQUEST => $this->createRequest('HEAD', 'http://example.com/'),
-                        HttpTransaction::KEY_RESPONSE => $this->createResponse(301),
-                        HttpTransaction::KEY_ERROR => null,
-                        HttpTransaction::KEY_OPTIONS => []
-                    ],
-                    [
-                        HttpTransaction::KEY_REQUEST => $this->createRequest('HEAD', 'http://example.com/1'),
-                        HttpTransaction::KEY_RESPONSE => $this->createResponse(301),
-                        HttpTransaction::KEY_ERROR => null,
-                        HttpTransaction::KEY_OPTIONS => []
-                    ],
-                    [
-                        HttpTransaction::KEY_REQUEST => $this->createRequest('HEAD', 'http://example.com/'),
-                        HttpTransaction::KEY_RESPONSE => $this->createResponse(301),
-                        HttpTransaction::KEY_ERROR => null,
-                        HttpTransaction::KEY_OPTIONS => []
-                    ],
-                    [
-                        HttpTransaction::KEY_REQUEST => $this->createRequest('GET', 'http://example.com/'),
-                        HttpTransaction::KEY_RESPONSE => $this->createResponse(301),
-                        HttpTransaction::KEY_ERROR => null,
-                        HttpTransaction::KEY_OPTIONS => []
-                    ],
-                    [
-                        HttpTransaction::KEY_REQUEST => $this->createRequest('GET', 'http://example.com/1'),
-                        HttpTransaction::KEY_RESPONSE => $this->createResponse(301),
-                        HttpTransaction::KEY_ERROR => null,
-                        HttpTransaction::KEY_OPTIONS => []
-                    ],
-                    [
-                        HttpTransaction::KEY_REQUEST => $this->createRequest('GET', 'http://example.com/'),
-                        HttpTransaction::KEY_RESPONSE => $this->createResponse(301),
-                        HttpTransaction::KEY_ERROR => null,
-                        HttpTransaction::KEY_OPTIONS => []
-                    ],
-                ],
-                'expectedHasRedirectLoop' => true,
-            ],
-            'redirecting indirectly back to self(with method group change, loop in first group only)' => [
-                'httpTransactions' => [
-                    [
-                        HttpTransaction::KEY_REQUEST => $this->createRequest('HEAD', 'http://example.com/'),
-                        HttpTransaction::KEY_RESPONSE => $this->createResponse(301),
-                        HttpTransaction::KEY_ERROR => null,
-                        HttpTransaction::KEY_OPTIONS => []
-                    ],
-                    [
-                        HttpTransaction::KEY_REQUEST => $this->createRequest('HEAD', 'http://example.com/1'),
-                        HttpTransaction::KEY_RESPONSE => $this->createResponse(301),
-                        HttpTransaction::KEY_ERROR => null,
-                        HttpTransaction::KEY_OPTIONS => []
-                    ],
-                    [
-                        HttpTransaction::KEY_REQUEST => $this->createRequest('HEAD', 'http://example.com/'),
-                        HttpTransaction::KEY_RESPONSE => $this->createResponse(301),
-                        HttpTransaction::KEY_ERROR => null,
-                        HttpTransaction::KEY_OPTIONS => []
-                    ],
-                    [
-                        HttpTransaction::KEY_REQUEST => $this->createRequest('GET', 'http://example.com/'),
-                        HttpTransaction::KEY_RESPONSE => $this->createResponse(301),
-                        HttpTransaction::KEY_ERROR => null,
-                        HttpTransaction::KEY_OPTIONS => []
-                    ],
-                    [
-                        HttpTransaction::KEY_REQUEST => $this->createRequest('GET', 'http://example.com/1'),
-                        HttpTransaction::KEY_RESPONSE => $this->createResponse(301),
-                        HttpTransaction::KEY_ERROR => null,
-                        HttpTransaction::KEY_OPTIONS => []
-                    ],
-                    [
-                        HttpTransaction::KEY_REQUEST => $this->createRequest('GET', 'http://example.com/2'),
-                        HttpTransaction::KEY_RESPONSE => $this->createResponse(301),
-                        HttpTransaction::KEY_ERROR => null,
-                        HttpTransaction::KEY_OPTIONS => []
-                    ],
-                ],
-                'expectedHasRedirectLoop' => true,
-            ],
-            'redirecting indirectly back to self(with method group change, loop in second group only)' => [
-                'httpTransactions' => [
-                    [
-                        HttpTransaction::KEY_REQUEST => $this->createRequest('HEAD', 'http://example.com/'),
-                        HttpTransaction::KEY_RESPONSE => $this->createResponse(301),
-                        HttpTransaction::KEY_ERROR => null,
-                        HttpTransaction::KEY_OPTIONS => []
-                    ],
-                    [
-                        HttpTransaction::KEY_REQUEST => $this->createRequest('HEAD', 'http://example.com/1'),
-                        HttpTransaction::KEY_RESPONSE => $this->createResponse(301),
-                        HttpTransaction::KEY_ERROR => null,
-                        HttpTransaction::KEY_OPTIONS => []
-                    ],
-                    [
-                        HttpTransaction::KEY_REQUEST => $this->createRequest('HEAD', 'http://example.com/2'),
-                        HttpTransaction::KEY_RESPONSE => $this->createResponse(301),
-                        HttpTransaction::KEY_ERROR => null,
-                        HttpTransaction::KEY_OPTIONS => []
-                    ],
-                    [
-                        HttpTransaction::KEY_REQUEST => $this->createRequest('GET', 'http://example.com/'),
-                        HttpTransaction::KEY_RESPONSE => $this->createResponse(301),
-                        HttpTransaction::KEY_ERROR => null,
-                        HttpTransaction::KEY_OPTIONS => []
-                    ],
-                    [
-                        HttpTransaction::KEY_REQUEST => $this->createRequest('GET', 'http://example.com/1'),
-                        HttpTransaction::KEY_RESPONSE => $this->createResponse(301),
-                        HttpTransaction::KEY_ERROR => null,
-                        HttpTransaction::KEY_OPTIONS => []
-                    ],
-                    [
-                        HttpTransaction::KEY_REQUEST => $this->createRequest('GET', 'http://example.com/'),
-                        HttpTransaction::KEY_RESPONSE => $this->createResponse(301),
-                        HttpTransaction::KEY_ERROR => null,
-                        HttpTransaction::KEY_OPTIONS => []
-                    ],
-                ],
-                'expectedHasRedirectLoop' => true,
-            ],
-        ];
-    }
-
     public function testGetTransactions(): void
     {
         $this->assertEquals([], $this->container->getTransactions());
-    }
-
-    private function createResponse(int $statusCode): ResponseInterface
-    {
-        $response = \Mockery::mock(ResponseInterface::class);
-        $response
-            ->shouldReceive('getStatusCode')
-            ->andReturn($statusCode);
-
-        return $response;
-    }
-
-    private function createRequest(string $method, string $url): RequestInterface
-    {
-        $uri = \Mockery::mock(UriInterface::class);
-        $uri
-            ->shouldReceive('__toString')
-            ->andReturn($url);
-
-        $request = \Mockery::mock(RequestInterface::class);
-
-        $request
-            ->shouldReceive('getMethod')
-            ->andReturn($method);
-
-        $request
-            ->shouldReceive('getUri')
-            ->andReturn($uri);
-
-        return $request;
     }
 }
